Apply converters after parsing
argparse's `type` gets applied before testing for choices. This makes it
unergonomic when trying to create an argument that'd accept a dict key
and return a value.

Using `type=d.__getitem__, choices=d.keys()` would not work, because
this would accept only dict *values* and not keys. The only possible
solution (I think) is to not provide a type and just do the conversion
after parsing is done.